InheritValidationFrom Validates Requests
Previously, we only pulled in the rules for the schema generation. Now, it will also validate the incoming request with the same rules its inheriting. There’s a 3rd flag on the InheritValidationFrom attribute now if you want to set it to ‘false’ and only use inheritance for the schema generation.

By default, it’s on, as this was my original intent and the most logical desire for developer experience.